% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_basic.R
\name{compile_basic}
\alias{compile_basic}
\title{Compile Basic Summaries}
\usage{
compile_basic(my_basic_summary_metals, my_basic_summary_other)
}
\arguments{
\item{my_basic_summary_metals}{obtained from
\code{\link{summarize_basic_metals}}.}

\item{my_basic_summary_other}{obtained from \code{\link{summarize_basic}}.}
}
\value{
dataframe with same columns as inputs. Column order may change.
}
\description{
Bind together metal and organic basic data summaries.
}
\examples{
try( compile_basic(my_basic_summary_other, my_basic_summary_metals) )
}
