% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_basic_recent.R
\name{compile_basic_recent}
\alias{compile_basic_recent}
\title{Compile Basic Summaries for Recent Timeframe}
\usage{
compile_basic_recent(
  my_basic_summary_recent_metals,
  my_basic_summary_recent_other
)
}
\arguments{
\item{my_basic_summary_recent_metals}{obtained from
\code{\link{summarize_basic_recent_metals}}.}

\item{my_basic_summary_recent_other}{obtained from
\code{\link{summarize_basic_recent}}.}
}
\value{
dataframe with same columns as inputs.
}
\description{
Bind together metal and organic basic data summaries for the same recent
timeframe.
}
\examples{
try( compile_basic_recent(my_basic_summary_10yr_metals, my_basic_summary_10yr_other) )
}
