% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_summaries.R
\name{compile_summaries}
\alias{compile_summaries}
\title{Compile All Summaries}
\usage{
compile_summaries(
  my_basic_summary,
  my_basic_summary_10yr,
  my_basic_summary_5yr
)
}
\arguments{
\item{my_basic_summary}{obtained from
\code{\link{compile_basic}}.}

\item{my_basic_summary_10yr}{obtained from
\code{\link{compile_basic_recent}}.}

\item{my_basic_summary_5yr}{obtained from
\code{\link{compile_basic_recent}}.}
}
\value{
dataframe with columns from inputs plus additional lookup tables.
}
\description{
Join basic summary with period based summaries. Add data from lookup tables.
}
\examples{
compile_summaries(my_basic_summary,
                  my_basic_summary_10yr,
                  my_basic_summary_5yr)
}
