% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ir_appendix_b_class_c.R
\name{create_ir_appendix_b_class_c}
\alias{create_ir_appendix_b_class_c}
\title{Create Class C Results Appendix Table}
\usage{
create_ir_appendix_b_class_c(my_decision_logic, five_year_start_date)
}
\arguments{
\item{my_decision_logic}{obtained from
\code{\link{create_decision_logic_class_c}}.}

\item{five_year_start_date}{beginning of date range for last five year
analysis in "mm/dd/yyyy" format.}
}
\value{
dataframe with formatting adjusted and columns renamed.
}
\description{
Format table for inclusion in report appendices.
}
\examples{
create_ir_appendix_b_class_c(my_decision_logic_class_c,
                             five_year_start_date = "07/01/2016")
}
