% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_results_reconciliation.R
\name{create_ir_appendix_b_reconciliation}
\alias{create_ir_appendix_b_reconciliation}
\title{Create Decision Reconciliation Table}
\usage{
create_ir_appendix_b_reconciliation(
  my_decision_logic_class_c,
  my_decision_logic_class_d
)
}
\arguments{
\item{my_decision_logic_class_c}{obtained from
\code{\link{create_decision_logic_class_c}}.}

\item{my_decision_logic_class_d}{obtained from
\code{\link{create_decision_logic_class_d}}.}
}
\value{
dataframe with Class C and Class D decisions and descriptions plus
the reconciled decision descriptions.
}
\description{
Create table for reconciliation between Class C and Class D decisions.
}
\examples{
try( create_ir_appendix_b_reconciliation(my_decision_logic_class_c,
                                          my_decision_logic_class_d) )
}
