% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_results_class_d.R
\name{create_results_class_d}
\alias{create_results_class_d}
\title{Create Class D Results Table}
\usage{
create_results_class_d(my_decision_logic, five_year_start_date)
}
\arguments{
\item{my_decision_logic}{obtained from
\code{\link{create_decision_logic_class_d}}.}

\item{five_year_start_date}{beginning of date range for last five year
analysis in "mm/dd/yyyy" format.}
}
\value{
dataframe with formatting adjusted and columns renamed.
}
\description{
Format table for inclusion in report appendices.
}
\examples{
try( create_results_class_c(my_decision_logic_class_d,
                            five_year_start_date = "07/01/2016") )
}
