% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_criteria.R
\name{evaluate_criteria}
\alias{evaluate_criteria}
\title{Evaluate Criteria}
\usage{
evaluate_criteria(df)
}
\arguments{
\item{df}{obtained from \code{\link{lookup_criteria}}.}
}
\value{
dataframe with boolean columns `exceedance_ccc`, `exceedance_cmc`,
  `exceedance_d`.
}
\description{
Determine whether Class C and Class D criteria are exceeded for each sample.
}
\examples{
try( evaluate_criteria(ir_data) )
}
