% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRsupportR-package.R
\docType{package}
\name{IRsupportR-package}
\alias{IRsupportR}
\alias{IRsupportR-package}
\title{IRsupportR: Analyze Toxics Data from District of Columbia Waterbodies}
\description{
The objective of this package is to compile and evaluate historic water quality data to verify if the data supports impairments attributed to toxics. In order to accomplish this objective, data from various datasets are compiled and processed into one consistent dataset for comparing to numeric water quality criteria. Exceedances of the water quality criteria are determined and categorization reevaluation decisions are proposed based on the implementation of an approved decision logic.
}
\author{
\strong{Maintainer}: Shanna Rucker \email{srucker@limno.com}

Authors:
\itemize{
  \item Ben Crary \email{bcrary@hazenandsawyer.com}
}

Other contributors:
\itemize{
  \item Diana Teran \email{dteran@limno.com} [contributor]
}

}
\keyword{internal}
