% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_criteria.R
\name{lookup_criteria}
\alias{lookup_criteria}
\title{Look Up Criteria}
\usage{
lookup_criteria(df)
}
\arguments{
\item{df}{obtained from all_processed_data.R script in data-raw folder.}
}
\value{
dataframe with water quality criteria columns added.
}
\description{
Add water quality criteria to dataset.
}
\examples{
try( lookup_criteria(ir_data) )
}
