% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_basic.R
\name{summarize_basic}
\alias{summarize_basic}
\title{Basic Summary of Organic Parameters}
\usage{
summarize_basic(criteria_results)
}
\arguments{
\item{criteria_results}{obtained from \code{\link{evaluate_criteria}}.}
}
\value{
dataframe with columns `most_recent_ccc_exceedance_date`,
  `most_recent_ccc_exceedance_year`, `most_recent_cmc_exceedance_date`,
  `most_recent_cmc_exceedance_year`, `most_recent_d_exceedance_date`,
  `most_recent_d_exceedance_year`, `n_since_most_recent_ccc_exceedance`,
  `n_since_most_recent_cmc_exceedance`, `n_since_most_recent_d_exceedance`.
}
\description{
Create basic summary statistics for all organic parameters.
}
\examples{
summarize_basic(criteria_results)
}
