% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_basic_metals.R
\name{summarize_basic_metals}
\alias{summarize_basic_metals}
\title{Basic Summary of Metals}
\usage{
summarize_basic_metals(criteria_results)
}
\arguments{
\item{criteria_results}{obtained from \code{\link{evaluate_criteria}}.}
}
\value{
dataframe with added columns `most_recent_ccc_exceedance_date`,
  `most_recent_ccc_exceedance_year`, `most_recent_cmc_exceedance_date`,
  `most_recent_cmc_exceedance_year`, `most_recent_d_exceedance_date`,
  `most_recent_d_exceedance_year`, `n_since_most_recent_ccc_exceedance`,
  `n_since_most_recent_cmc_exceedance`, `n_since_most_recent_d_exceedance`.
}
\description{
Create basic summary statistics for all metal parameters.
For metals, sample counts depend on the test fraction. Class C uses dissolved
metals. Class D uses total metals.
}
\examples{
try( summarize_basic_metals(criteria_results) )
}
