% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_basic_recent.R
\name{summarize_basic_recent}
\alias{summarize_basic_recent}
\title{Basic Summary of Organic Parameters for Recent Timeframe}
\usage{
summarize_basic_recent(criteria_results, start_date, end_date)
}
\arguments{
\item{criteria_results}{obtained from \code{\link{evaluate_criteria}}.}

\item{start_date}{beginning of date range of interest in "mm/dd/yyyy" format.}

\item{end_date}{end of date range of interest in "mm/dd/yyyy" format.}
}
\value{
dataframe with added columns based on specified timeframe.
}
\description{
Create basic summary statistics for all organic parameters within specified
timeframe.
}
\examples{
summarize_basic_recent(criteria_results = criteria_results,
                       start_date = "07/01/2011",
                       end_date = "06/30/2021")
}
