% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ir_appendix_b_class_c_merge.R
\name{create_ir_appendix_b_class_c_merge}
\alias{create_ir_appendix_b_class_c_merge}
\title{create_ir_appendix_b_class_c}
\usage{
create_ir_appendix_b_class_c_merge(my_decision_logic, five_year_start_date)
}
\arguments{
\item{df}{}
}
\description{
create_ir_appendix_b_class_c
}
